move file row into separate component
